Get panes to show correct view.
Trying to make a system where the user can create multiple
panes. Each pane can be certain type of view, for example
a Buffer view, a Texture view, a Step view, etc....

TODO: Make the bar above each view have icons for splitting
and changing the view type. I plan to leave this to the end
since it's not important to get basic functionality working.