eslint: turn off no-undef rule for samples
The samples are in JavaScript and eslint doesn't know about WebGPU
so lint complains about all of the webgpu types.

We could those types as globals in the .eslintrc file or maybe
there is some other way to make eslint aware of those types
but for now I just turned off that rule for the samples.