Add code to give you an effectively, unwrapped GPUAdapter
You call `requestUnwrappedAdaptor(options)` and you'll get
back an GPUAdaptor that functions as thought it's unwrapped.
This works by copying the original functions as properties
on the GPUAdaptor from requestAdaptor so those will be used
instead of the functions on the prototype. Functions that
recreate objects, for example requestDevice, createBuffer, etc
are wrapped so this same process is applied to objects they
create.

This way, the user's code will still be tracked for capture
but the Replay code and the UI code and use the API.

There's also `requestUnwrappedWebGPUContext` for getting a
GPUCanvasContext from an HTMLCanvasElement.

Note: I moved the current test playback code into
src/samples/traceFrame.js so it can be updated in one place.