Switch to FlexLayout from react-tile-pane
Some things I couldn't get to work with react-tile-pane
so switching to FlexLayout. Also, couldn't get FlexLayout
to build because of some strange dependency issue. It was
easy to just include its source to work around that issue
so doing that for now.

If we find a solution that lets us keep it as a package then
we'll switch.