Hack in a 'replayCount'
I'm not sure if there is a better way to do this. The issue is that
React will not re-render (in this case the `useEffect` function) unless
one if it's dependencies changes. The texture being passed, even though
the texture's content has changed, the texture itself has not. It's the
same object so `previousInternallyCachedValue === texture` returns true
and react doesn't run the effect.

So, I passed in a 'replayCount' that increments each time the user
clicks a step. That way react will see a change and things get
re-rendered