Get Steps to try/catch using JSON.stringify on complex objects
This a temporary fix. It's not clear how to go from the `command.args`
to what to display in steps. If it's an object should it just show
the name of the argument and then click to inspect elsewhere?

For now, it's using JSON.stringify but if that fails it shows an error.