make npm run check fail on typescript warnings
I couldn't figure out how to suppress the errors from FlexLayout
so as a workaround I run the build and capture its output and try
to check for typescript line number errors in the output.

The FlexLayout errors don't have a line number so this hacked workaround
seems to work for now.