Add Load and Save
It's a little concerning that a simple textured cube is 5mb!

Not sure what to do about that. Even if we compress the JSON with like
zip it will still need to be decompressed to a string and parsed.

I'm guessing to do handle heavier apps will require putting the resource
in binary formats, maybe like gLTF where all the buffers and textures
are in one or more separate files.