Make some panes not closable.
This solution is probably more complicated than it needs to be and is
probably complete.

The current design, influenced by the previous UI library, is that
the UI library is just displaying <Pane> components and it knows nothing
else. What to display in a Pane component is associated by the name of
the pane of registered in the UI library. So <Pane> just needs to know
the name it's using and then look up in unrelated state what Component
to display with what data.

So, along comes the issue that based on the Component type, the close
button should appear or not. The solution here is 2 fold

1. At init time, set enableClose for each tab by searching through the
   Json layout description by ids.

2. At setPaneViewType time. Call a function the UI registered
   that can then be used to call into the UI library and set the
   enableClose flag for a given tab.