Serialize a "swapChainId" for TextureState
The issue is that 100s of textures from `getCurrentTexture` end up in
in UI as just `GPUTexture` and it's hard to tell them apart.

So, this tacks on the `id` of the texture and the call count for that
GPUCanvasContext to the texture state. They then show up in the UI as

```
GPUtexture:label:[gct myCanvas-123]
GPUtexture:label:[gct myCanvas-124]
GPUtexture:label:[gct myCanvas-125]
GPUtexture:label:[gct myCanvas-126]
```

etc.