Fix some shader and table syntax issues
Changes to Tint meant that the texture vis and picker code was
only working in Canary. Made some tweaks so that it'll work in dev
in the meantime. Also fixed an issue that Rect was yelling at me
about in terms of table formatting. Don't know why it wasn't
showing me the same error earlier.