Add more types
I'm not sure I'm happy about this change. There were JavaScript-isms
that don't really seem to port to TypeScript. For example:

The trace code would generate a 'present' command in JS as

```js
traceCommand({
  name: 'present',
  args: {
    textureSerial: getSerialForTexture(..);
  }
});
```

The replay code would then hack that

```js
const c = command[ndx++];
c.args.texture = this.textures[c.args.textureSerial];
delete c.args.textureSerial;
```

I typed the trace command as

```ts
interface TraceQueueCommandPresent {
  name: 'present'
  args: {
    textureSerial: number;
  };
}
```

I can't really add `texture?: ReplayTexture` there as the tracing
code should have no connection to the replay code.

But, giving the replay code it's own types means that each type has
to be manually copied, field by field from a TraceQueueCommandPresent
to a ReplayQueueCommandPresent (or not? see below)

That's fine I suppose, it should at least be able to check you filled
out all the required fields. It was just convenient to be able to copy
the objects and get all the data. As it is now, if you add a field to
the trace type you need to add code to copy that field to the
corresponding replay type. I don't know that that's a big deal as if you're adding
the field then presumably you want to use it in the replay so you're going
to add it to the replay type. And besides, you probably want the type
documented. It's mostly just that the code expanded a bunch by having
to declare all the types twice (trace vs replay) and add the code to
translate from trace -> replay.

Also, I'm not sure it matters but is there any point in making the
commands consistent. Some commands have `args`. Other commands just
put they data at the root of the command.

On the UI side, it would be nice, where possible, to be given some
property like `args` and the UI would just display the all. As it is
now the UI will have to add code per type to try to display things
correctly.

Maybe working on it more it will become clearer if there is a more
generic way to organize the data.