100% Typescript Typed
Well, I guess 99%. The code that converts queue commands is marshelling
commands but the code that converts command buffer commands is still
using `window.structuredClone` and then adding and removing fields.
That code is not typesafe but it works for now.

It's not clear to me which is better. The good part is if you add
a field upstream it gets added/copied downstream. The bad part is
you can't really type it in that spot.